+++ subsys/bluetooth/controller/ll_sw/lll_adv_aux.h
