+++ subsys/bluetooth/controller/ll_sw/lll_adv_sync.h
