+++ subsys/bluetooth/controller/ll_sw/lll_master.h
