+++ subsys/bluetooth/controller/ll_sw/lll_scan_aux.h
