--- subsys/bluetooth/controller/ll_sw/openisa/lll/lll_slave.h
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2018-2019 Nordic Semiconductor ASA
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-int lll_slave_init(void);
-int lll_slave_reset(void);
-void lll_slave_prepare(void *param);
