+++ subsys/bluetooth/controller/ll_sw/ull_filter.c
@@ -27,6 +27,7 @@
 #include "lll_scan.h"
 #include "lll_conn.h"
 #include "lll_filter.h"
+#include "lll_adv_internal.h"
 
 #include "ull_adv_types.h"
 #include "ull_scan_types.h"
